big: implement the core business logic in flutter (cloud)
The idea is to move most of the separate android/ios v6 codebases
into one language, and hook it up to the native platforms for
the necessary calls. Most code should be in one dart codebase,
easy to test and well covered.

This commit contains all core and cloud logic (no plus logic yet).