vpn: improve turnOn retry logic
There is some problem with either the iOS WG lib itself, or our
platform integration with it in Swift, but apparently sometimes
the VPN established won't have connecitivity. As far as I can
tell, there is no difference between the "good" and "bad"
scenarios in terms of how we call the WG. The only difference
that I noticed is that in the bad state, WG won't receive
handshake response, and also there is some network path
related case "Connectivity offline" in the Console log.

The workaround here is to just re-attempt establishing VPN
several times in a row in case if the first lease request
made just after establishing the VPN, fails. It acts as a
health check for the newly established VPN. In practice,
this seems to help and the VPN gets established after
the retry (didn't see it need to retry more than once so far,
but the limit is 5).